Improvements, formatting
